Switch metrics from usize to u64
It's easier to convert usize values to u64 values than it is to go the
other way. Metrics are generally used as outputs rather than inputs, so
it makes sense to have it be the type that everything can easily convert
to.

+++ src/vmm/src/devices/legacy/i8042.rs
@@ -205,7 +205,7 @@ impl I8042Device {
             _ => read_ok = false,
         }
         if read_ok {
-            METRICS.i8042.read_count.add(data.len());
+            METRICS.i8042.read_count.add(data.len() as u64);
         } else {
             METRICS.i8042.missed_read_count.inc();
         }
